// src/generated/prisma/enums.ts
export const JournalType = {
  SALES: 'SALES',
  PURCHASE: 'PURCHASE',
  BANK: 'BANK',
  CASH: 'CASH'
} as const

export type JournalType = (typeof JournalType)[keyof typeof JournalType]


export const EntryState = {
  DRAFT: 'DRAFT',
  POSTED: 'POSTED',
  CANCELLED: 'CANCELLED'
} as const

export type EntryState = (typeof EntryState)[keyof typeof EntryState]


export const EntrySource = {
  MANUAL: 'MANUAL',
  VENDOR_BILL: 'VENDOR_BILL',
  CUSTOMER_INVOICE: 'CUSTOMER_INVOICE',
  PAYMENT: 'PAYMENT',
  STOCK_MOVE: 'STOCK_MOVE',
  REVERSAL: 'REVERSAL',
  OPENING_BALANCE: 'OPENING_BALANCE'
} as const

export type EntrySource = (typeof EntrySource)[keyof typeof EntrySource]